update for ghc-8.4.3
The main change is mainly for upgrade of extensible from v0.4.7 to v0.4.10.
Since v0.4.8, extensible's `Record` implements some type classes including
`ToJSON` and `FromJSON`. To cope with this, `Record` of nvim-hs-lsp now
got a newtype of extensible's `Record`. The original `Record` is defined
in `Data.Extensible.Rexport` as `OrigRecord`.